+++ llm_studio/src/models/text_causal_language_modeling_model.py
@@ -8,7 +8,7 @@ from transformers import AutoModelForCausalLM, StoppingCriteria, StoppingCriteri
 from transformers.generation.utils import GenerationMixin
 from transformers.utils import logging as transformers_logging
 
-from llm_studio.src.losses.classification_losses import Perplexity
+from llm_studio.src.metrics.text_causal_language_modeling_metrics import Perplexity
 from llm_studio.src.utils.data_utils import batch_padding
 from llm_studio.src.utils.modeling_utils import create_nlp_backbone
 
@@ -120,7 +120,7 @@ class Model(nn.Module):
         )(self.cfg)
 
         if self.cfg.prediction.metric == "Perplexity":
-            self.perplexity = Perplexity(self.cfg)
+            self.perplexity = Perplexity(self.cfg, reduce=False)
 
     def generate(self, batch: Dict, cfg: Any):
         pad_token_id = (
@@ -203,9 +203,10 @@ class Model(nn.Module):
             if calculate_loss:
                 outputs["loss"] = self.loss_fn(output.logits, batch["labels"])
 
-        if self.cfg.prediction.metric == "Perplexity":
-            outputs["perplexity"] = self.perplexity(output.logits, batch["labels"])
-        elif not self.training:
+            if self.cfg.prediction.metric == "Perplexity":
+                outputs["perplexity"] = self.perplexity(output.logits, batch["labels"])
+
+        if not self.training and self.cfg.prediction.metric != "Perplexity":
             outputs["predicted_answer_ids"] = self.generate(batch, self.cfg)
 
         return outputs
